extractor/requirements.txt
requests>=2.31.0
requests-oauthlib>=1.3.1
python-dotenv>=1.0.0
supabase>=2.4.0
pydantic>=2.6.0
deepdiff>=6.7.0
